+++ config/routes.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 RecordingStudioPublications::Engine.routes.draw do
-  root "home#index"
+  namespace :admin do
+    resources :publications, only: %i[index new create show edit update]
+  end
 end