+++ test/dummy/config/routes.rb
@@ -6,6 +6,8 @@ Rails.application.routes.draw do
   get "/recording_studio", to: redirect("/"), as: nil
   mount RecordingStudio::Engine, at: "/recording_studio"
   mount RecordingStudioRootSwitchable::Engine, at: "/recording_studio_root_switchable"
+  mount RecordingStudioPublications::Engine, at: "/recording_studio_publications"
+  recording_studio_admin_for :admin, at: "/admin", root_section: :publications
 
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html